% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_bayesest.R
\name{benchmark_bayesest}
\alias{benchmark_bayesest}
\title{Benchmark posterior draws via a constrained Bayes estimator approach}
\usage{
benchmark_bayesest(
  posterior_samps,
  nregion = NA,
  ntime = 1,
  natl,
  pop_weights = NA
)
}
\arguments{
\item{posterior_samps}{a matrix containing posterior draws at the area-level. Rows should be
arranged in order arrange(region, time), with each column containing an independent draw from
the posterior.}

\item{nregion}{the number of regions you have estimates for. Defaults to the number of rows
in \code{posterior_samps}.}

\item{ntime}{the number of time points you have estimates for. Defaults to 1.}

\item{natl}{a vector of national level estimates, arranged in order of time if you have multiple
time points in your data.}

\item{pop_weights}{a vector of population weights for use in the benchmarking constraint. Must
sum to one at each time point, and be in order arrange(region, time)}
}
\value{
A matrix of benchmarked samples in order arrange(region, time)
}
\description{
Benchmarks draws from a distribution using the method described in
the paper "Bayesian benchmarking with applications to small area estimation" by
G.S. Datta, M. Ghosh, R. Steorts, and J. Maples. Takes in a matrix of posterior draws,
and national-level benchmark data.
}
\author{
Taylor Okonek
}
