% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_mh.R
\name{benchmark_mh}
\alias{benchmark_mh}
\title{Benchmark posterior draws via a Metropolis-Hastings algorithm}
\usage{
benchmark_mh(
  posterior_samps,
  posterior_samps_fe,
  nregion = NA,
  ntime = 1,
  natl,
  natl_sd,
  pop_weights = NA,
  fe_prior_means,
  fe_prior_variances
)
}
\arguments{
\item{posterior_samps}{a matrix containing posterior draws at the area-level. Rows should be
arranged in order arrange(region, time), with each column containing an independent draw from
the posterior.}

\item{posterior_samps_fe}{a matrix containing posterior draws of fixed effects (typically intercepts)
that have had prior means and variances shifted in the unbenchmarked model, with each column
containing an independent draw from the posterior.}

\item{nregion}{the number of regions you have estimates for. Defaults to the number of rows
in \code{posterior_samps}.}

\item{ntime}{the number of time points you have estimates for. Defaults to 1.}

\item{natl}{a vector of national level estimates, arranged in order of time if you have multiple
time points in your data.}

\item{natl_sd}{a vector of standard deviations for national level estimates, arranged in order of
time if you have multiple time points in your data.}

\item{pop_weights}{a vector of population weights for use in the benchmarking constraint. Must
sum to one at each time point, and be in order arrange(region, time)}

\item{fe_prior_means}{a vector of prior means for the fixed effects specified in \code{posterior_samps_fe}. Must
be in row order of posterior_samps_fe.}

\item{fe_prior_variances}{a vector of prior variances for the fixed effects specified in \code{posterior_samps_fe}
Must be in row order of posterior_samps_fe.}
}
\value{
A list containing:
\itemize{
\item fitted_list: a list of matrices of benchmarked posterior samples of fitted values in
order arrange(time). Each matrix will have rows arranged in order arrange(region).
\item natl_list: a list of vectors containing aggregated national-level samples that were
accepted, in order arrange(time)
\item prop_accepted: the proportion of samples accepted during sampling.
}
}
\description{
Benchmarks draws from a posterior distribution using a Metropolis-Hastings approach. Takes in
a matrix of posterior draws, a matrix of fixed effect draws, and national-level benchmark data.
}
\author{
Taylor Okonek
}
