% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_normal.R
\name{fit_normal}
\alias{fit_normal}
\title{Fit normal outcome model in TMB}
\usage{
fit_normal(
  df,
  cluster = NA,
  y = "y",
  region = "admin1",
  SE_i,
  hiv_adj = NA,
  natl = NULL,
  natl_sd = NULL,
  pop_weights = NULL,
  Q_struct_space = NULL,
  intercept_pri = c(0, 31.62278),
  nsamp = 1000,
  benched = "unbenched",
  expit_outcome = NULL
)
}
\arguments{
\item{df}{a dataframe containing normal outcome data, and the following columns:
\itemize{
\item \code{cluster}: cluster ID
\item \code{y}: normally distributed outcome
\item \code{region}: area
}}

\item{cluster}{the column in \code{binom_df} corresponding to cluster id. If data is area-level
then cluster should be set to NA. By default, cluster will be set to NA.}

\item{y}{the column in \code{binom_df} corresponding to the outcome}

\item{region}{the column in \code{binom_df} corresponding to region/area. Must be numeric and
start from 1.}

\item{SE_i}{a vector of fixed standard errors for use in the normal likelihood. Length of
\code{SE_i} must be equal to the number of regions.}

\item{hiv_adj}{An optional log offset in time to include in the linear predictor. Defaults to
no log offset.}

\item{natl}{a vector of national level estimates, arranged in order of time}

\item{natl_sd}{a vector of standard deviations for national level estimates, arranged in order of
time}

\item{pop_weights}{a vector of population weights for use in the benchmarking constraint. Must
sum to one at each time point, and be in order arrange(region)}

\item{Q_struct_space}{An ICAR precision matrix. Should be unscaled, as scaling will happen
internally.}

\item{intercept_pri}{Prior specification for the intercept. Defaults to c(0, 31.62278), corresponding
to the default prior for the intercept in INLA, with mean 0 and precision 0.001. Must be
a vector of length 2, with specificaiton c(mean, sd) for a Normal distribution. Currently only
an option for the area-level, unbenchmarked model.}

\item{nsamp}{Number of posterior samples to take from joint posterior. Defaults to 1000}

\item{benched}{A string, either \code{"benched"}, \code{"unbenched"}, or \code{"both"}, determining
whether to fit a benchmarked model, and unbenchmarked model, or both. Defaults to \code{"unbenched"}.}

\item{expit_outcome}{A boolean for whether or not the normally distributed outcome needs to be
expit-ed in order to be on the same scale as national estimates. This parameter will only affect
benchmarked models. If TRUE, linear predictors will be expit-ed before being multiplied by population
weights in the benchmarking constraint. This parameter must be specified by the user if
\code{benched = "benched"}.}
}
\value{
A list containing:
\itemize{
\item fitted_mat: a matrix of posterior samples of fitted values in order arrange(region, time)
\item re_list: a list contaning matrices of posterior samples for each random effect term
\item param_list: a list containing matrices of posterior samples for fixed effects and hyperparameters
\item runtime: the time it took to fit the model in TMB and get samples from the joint posterior
}
If \code{benched = "both"}, a list of two will be returned containing the above list for both
benchmarked and unbenchmarked models.
}
\description{
Fit normal outcome model in TMB. Currently can only fit a space-only model
at a single time point. Linear predictor includes a single intercept and
a BYM2 sptial random effect, and an IID cluster-level random effect is included
if unit-level data is input (otherwise, no cluster-level random effect).
Can fit benchmarked models, unbenchmarked models, or both.
Simultaneous benchmarking is performed via a second likelihood for the national level estimate.
Produces posterior samples for fitted values, hyperparameters,
random effects, and fixed effects.
}
\author{
Taylor Okonek
}
