% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stbench-package.R
\docType{package}
\name{stbench-package}
\alias{stbench}
\alias{stbench-package}
\title{stbench: Fully Bayesian Benchmarking for spatio-temporal models}
\description{
An open-source R package for producing fully Bayesian benchmarked estimates for hierarchical models in TMB. The package currently supports spatial and spatio-temporal modeling of binomial count data of under-5 mortality (U5MR), spatial models for generic binomial count data, and spatial models for normally distributed outcomes.
}
\author{
\strong{Maintainer}: Taylor Okonek \email{tokonek@uw.edu}

}
\keyword{internal}
